+++ src/main/java/app/models/entities/Pagamento.java
@@ -12,7 +12,7 @@ public class Pagamento {
     private final Integer id;
     private final String cpf;
     private final Integer idFonte;
-    private final Integer numeroContrato;
+    private final Long numeroContrato;
     private final Date dataVencimentoParcela;
     private final Date dataPagamentoParcela;
     private final BigDecimal valorPago;
